Limit the work to data considered "interesting"
A loan is interesting if it's one for which an error can occur.
An origin is interesting if an interesting loan can flow into it, or if it's a placeholder origin (though this can be further improved).
A `borrow_region` is interesting if its loan is interesting.
An `outlives` constraint is interesting if an interesting origin flows into another origin.